refactor : "DetailView" 에서 사진 탭을 인식하는 메서드 구현
- UITapGestureRecognizer 를 이용.
- 사진 탭시, 원본비율을 볼 수 있는 화면으로 전환하는 기능을 구현할 예정.
