--- YagomMarket/YagomMarket/Views/DetailView.swift
@@ -1,151 +1,0 @@
-//
-//  DetailView.swift
-//  YagomMarket
-//
-//  Created by 이원빈 on 2022/11/28.
-//
-
-import UIKit
-
-final class DetailView: UIView {
-    // MARK: Properties
-    private let viewModel = DefaultDetailViewModel()
-    
-    // MARK: UIComponents
-    private let mainScrollView: UIScrollView = {
-        let scrollView = UIScrollView()
-        scrollView.translatesAutoresizingMaskIntoConstraints = false
-        return scrollView
-    }()
-    
-    private let mainStackView: UIStackView = {
-        let stackView = UIStackView()
-        stackView.translatesAutoresizingMaskIntoConstraints = false
-        stackView.isLayoutMarginsRelativeArrangement = true
-        stackView.layoutMargins = UIEdgeInsets(top: 20, left: 20, bottom: 0, right: 20)
-        stackView.axis = .vertical
-        stackView.spacing = 5
-        return stackView
-    }()
-    
-    private let imageView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.translatesAutoresizingMaskIntoConstraints = false
-        imageView.contentMode = .scaleAspectFill
-        imageView.clipsToBounds = true
-        return imageView
-    }()
-    
-    private let priceLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = UIFont.boldSystemFont(ofSize: 28)
-        return label
-    }()
-    
-    private let lowerStackView: UIStackView = {
-        let stackView = UIStackView()
-        stackView.axis = .horizontal
-        stackView.spacing = 8
-        return stackView
-    }()
-    
-    private let vendorNameLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.setContentHuggingPriority(.required, for: .horizontal)
-        label.textColor = .systemGray
-        return label
-    }()
-    
-    private let timeLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.textColor = .systemGray
-        return label
-    }()
-    
-    private let nameLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = UIFont.boldSystemFont(ofSize: 20)
-        return label
-    }()
-    
-    private let descriptionTextView: UITextView = {
-        let textView = UITextView()
-        textView.isScrollEnabled = false
-        textView.font = UIFont.preferredFont(forTextStyle: .body)
-        textView.translatesAutoresizingMaskIntoConstraints = false
-        return textView
-    }()
-    // 이 판매자가 올린 상품들을 밑에
-    
-    // MARK: Initializers
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        addSubViews()
-        setupConstraints()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func addSubViews() {
-        addSubview(mainScrollView)
-        
-        mainScrollView.addSubview(imageView) //
-        mainScrollView.addSubview(mainStackView)
-        
-//        mainStackView.addArrangedSubview(imageView)
-        mainStackView.addArrangedSubview(priceLabel)
-        mainStackView.addArrangedSubview(nameLabel)
-        mainStackView.addArrangedSubview(lowerStackView)
-        mainStackView.addArrangedSubview(descriptionTextView)
-        
-        lowerStackView.addArrangedSubview(vendorNameLabel)
-        lowerStackView.addArrangedSubview(timeLabel)
-    }
-    
-    private func setupConstraints() {
-        NSLayoutConstraint.activate([
-            mainScrollView.topAnchor.constraint(equalTo: topAnchor),
-            mainScrollView.bottomAnchor.constraint(equalTo: bottomAnchor),
-            mainScrollView.leadingAnchor.constraint(equalTo: leadingAnchor),
-            mainScrollView.trailingAnchor.constraint(equalTo: trailingAnchor),
-            
-            imageView.topAnchor.constraint(
-                equalTo: mainScrollView.contentLayoutGuide.topAnchor
-            ),
-            mainStackView.topAnchor.constraint(
-                equalTo: imageView.bottomAnchor
-            ),
-            mainStackView.bottomAnchor.constraint(
-                equalTo: mainScrollView.contentLayoutGuide.bottomAnchor
-            ),
-            mainStackView.leadingAnchor.constraint(
-                equalTo: mainScrollView.contentLayoutGuide.leadingAnchor
-            ),
-            mainStackView.trailingAnchor.constraint(
-                equalTo: mainScrollView.contentLayoutGuide.trailingAnchor
-            ),
-            mainStackView.widthAnchor.constraint(
-                equalTo: mainScrollView.widthAnchor
-            ),
-            imageView.heightAnchor.constraint(
-                equalToConstant: UIScreen.main.bounds.height * 0.5
-            ),
-            imageView.widthAnchor.constraint(equalToConstant: UIScreen.main.bounds.width)
-        ])
-    }
-    
-    func setupData(with viewModel: DetailViewModel) {
-        nameLabel.text = viewModel.name
-        descriptionTextView.text = viewModel.description
-        vendorNameLabel.text = viewModel.vendorName?.appending(" •")
-        priceLabel.text = viewModel.price
-        timeLabel.text = viewModel.time
-        imageView.setImage(with: (viewModel.images?[0].url)!)
-    }
-}
